Incoming multiple messages working
Client is able to receive multiple messages from server successfully.
Havent done any bug fixing or parsing of the message yet, but we’re
getting there. Server knows when to stop as well.  Currently theres a
bias towards client[0], but we’ll need further testing of this once it
actually does decryption to see if it fully works